# rlp: reduce allocations for big.Int and byte array encoding