This change further improves the performance of RLP encoding by removing
allocations for `big.Int` and `[...]byte` types. I have added a new benchmark
that measures RLP encoding of `types.Block` to verify that performance is
improved.

Here are the results as measured on my flimsy 1.2GHz Intel Core m3:

```
name           old time/op    new time/op    delta
EncodeBlock-4    68.6µs ± 6%    50.2µs ± 7%  -26.81%  (p=0.000 n=15+15)
name           old alloc/op   new alloc/op   delta
EncodeBlock-4    9.73kB ± 0%    0.06kB ± 1%  -99.34%  (p=0.000 n=11+15)
name           old allocs/op  new allocs/op  delta
EncodeBlock-4       465 ± 0%         1 ± 0%  -99.78%  (p=0.000 n=15+15)
```