@holiman 2020-07-05
Can confirm
```
name           old time/op    new time/op    delta
EncodeBlock-6    63.2µs ±16%    45.2µs ± 5%  -28.53%  (p=0.000 n=10+8)

name           old alloc/op   new alloc/op   delta
EncodeBlock-6    9.73kB ± 0%    0.07kB ± 0%  -99.32%  (p=0.000 n=10+8)

name           old allocs/op  new allocs/op  delta
EncodeBlock-6       465 ± 0%         1 ± 0%  -99.78%  (p=0.000 n=10+10)
```